nvmf/auth: check for support when adding dhchap keys
In-band authentication requires certain version of openssl (it needs to
define the EVP_MAC* functions).  With an older version, it's possible to
specify a dhchap key when adding a host, but it wouldn't be possible to
use it for authentication because calculating DH-HMAC-CHAP response
would always fail.

This behavior could be a bit misleading to the users, so this patch
makes it impossible to add a host with dhchap key if the version of
openssl is too old.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I03738b0c16d4b641aaa165262284cd6f5ca9a7aa
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22657
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Ben Walker <ben@nvidia.com>