test/iscsi: remove duplicate fuzzing test
Both autofuzz_iscsi.sh and fuzz.sh are the same.
The difference is passing the time to fuzz and
minor difference in how the application was started.

Removed fuzz.sh which was called directly in per-patch
autotest for short period of 30 sec. Replaced it with call
to autofuzz_iscsi.sh with the same timer.

Adjust ISCSI_APP to start in the same way, most importantly
per-patch now won't generate megabytes of logs due to
lack of output redirection in the removed fuzz.sh.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I79927a0634d9aa96014f9b589fd2883268ee18f5
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22518
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>