event: enforce 1 second default scheduling period
Currently the scheduling period was an optional parameter for
the framework_set_scheduler RPC. The 'dynamic' scheduler
enforced its own default by explicitly setting the period
to 1 second during init, if no value had been specified.

But other schedulers (gscheduler) didn't have that same code.
We could add it, but it's better to just enforce this when
the scheduler is first set - if user didn't pass a period,
set it to 1s.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ie2e1b3326438b34ebe0f3f65bf568e16621a288b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22593
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>