nvme/cuse: don't ERRLOG for NVME_IOCTL_ID in controller ioctl path
nvme-cli will often send this IOCTL to controller device nodes. For example,
some commands such as get_features *may* take a namespace ID, so nvme-cli
just sends this IOCTL to the device node to get the namespace ID if it exists.

So add a case for this ioctl cmd, so that we still return ENOTTY but without
the ERRLOG.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Id80ab6cf41da1783adb2a749f8dc492d075039a9
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22594
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>