trace: rename owner to owner_type
We will be adding the real concept of an "owner" in upcoming patches. So change
the existing "owner" to "owner_type" which more accurate reflects its meaning.
For example, an "owner_type" can be OWNER_BDEV, but an actual "owner" would
be some specific bdev identified by its name.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ieb25eb5a574a88b3812291f2345796e5bb66a54f
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22645
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>