nvmf/tcp: register and use trace owners
We now use the new trace owner functionality in the nvmf/tcp layer.

When a tcp qpair is initially created, we will register it as a trace owner,
with a description that just contains the host's IP/port. Later once the
CONNECT specifies the subsys, we add that and update the description. This
ensures we get meaningful events on a qpair even before the CONNECT is
established.

Now that we are using the owner_id, we no longer need to specify the qpair
pointer as an extra parameter to any of the nvmf/tcp trace events, so remove
them.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I7c01e061106b824ebbc5f12bd556806a1f5788fa
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22650
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>