app/trace: emit owner descriptions
Now that we have a real description for an event's owner, print those
descriptions instead of the old owner_type prefix + poller_id.

Note that we need to take the event's tsc into account, if it is older than
when the associated owner was registered we should not print the description
because it is for an owner that was unregistered after the event was
generated.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I9257306b5fe9fd7273abcb9b14e4fc136c417742
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22648
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>