raid: examine other bdevs when starting from superblock
When a bdev containing a superblock is examined, search for other base
bdevs and claim them, without having to examine every base bdev
separately.

For example, if a raid bdev is stopped and we remove one base bdev, then
add it back, this will now cause the raid bdev to be brought back up.
Earlier, we would have to re-add every base bdev.

Related to #3306

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Change-Id: I71193de1cdb8b3a451aa8490255d4414d81da159
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22641
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>