nvme/auth: make DH functions public
We'll want to use these functions on the target side too.

This patch also introduces struct spdk_nvme_dhchap_dhkey, which is only
used to avoid exposing openssl's EVP_PKEY on the interface.  It'll also
make it possible to add extra context to a DH key, if we ever need it,
without having to modify the API.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I67359b40c844fa1142ade5e35897ca46ace7cc61
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22757
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Seung yeon Shin <syeon.shin@samsung.com>