sock: replace sock impl priorities
Instead of using socket priorities, use a macro to set the default
implementation.
POSIX is the simplest impl compared to other sockets in SPDK and it
should be used as default (POSIX is most likely to be present at all
times and has the minimal requirements to work).

Change-Id: I8f0dc47777183154773ac6905aa9306418657e06
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22139
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>