sock: add default impl override
Read environmental variable 'SPDK_SOCK_IMPL_DEFAULT' and override the
default socket implementation (POSIX) with its value during
sock_subsystem_init().

This is done mainly to enable original behavior where a socket with
highest priority would be used by default (e.g. uring if configure
was run with '--with-uring' option). That way we may explicitly enable
chosen implementation as default, without relying on RPC interface.

As part of this patch, link sock modules into the app_repeat test app.
The method used here works even when for the app repeat case, so let's test it.
This allows us to just generally declare this flag for all tests and not have
to worry about apps like app_repeat that link in the sock library but none of
the sock modules.

Change-Id: I6af633cbe5bfba065580ba6c501169cbb5b64541
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22691
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>