sock/uring: make 'buf' 8-byte aligned
In spdk_uring_sock structure, a padding is needed to keep
'buf' aligned to 8-bytes. 'struct cmsghdr' is mapped
to this buffer in uring_sock_group_impl_add_sock() function,
and while first element of this control message header
has a size of 8 bytes, the whole structure must be
8-byte aligned, otherwise UBSAN will generate
runtime error.

This patch is part of the fix for github issue 3090.

Change-Id: I4f0b99b4cec23f541024f2d4bbcc2ab49db6ed05
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22741
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot