test/nvmf: relax subsystem filter in ns_masking.sh
`nvme list-subsys` returns an array of objects, one for each HostNQN. In some
cases, the first entry might be the OS's hostnqn, with an SSD subsystem. Then
the ns_masking.sh subsystem ends up in the second entry.

But the filter was explicitly only looking at the first entry. So relax
this to look at all entries.

Note that we could filter on the hostnqn, but that just adds more code to
the test script, filtering on subsystem NQN will always be sufficient.

Fixes issue #3337.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Idf7adfe6326d0468e576c190e7086a80213fe7eb
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22747
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>