bdev/nvme: make ROUND_ROBIN the default multipath policy for active/active
ROUND_ROBIN should be the default, but the patch that added the QUEUE_DEPTH
selector accidentally broke this.

The net effect would be that the selector would be set to 0 which is not
a valid enum value, but would end up acting like QUEUE_DEPTH selector since
the critical decision compares the selector with ROUND_ROBIN.

Fixes issue #3338.
Fixes: 6aa4edc ("bdev/nvme: select io path according to outstanding io numbder")
Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I930a9cf0f1078680854712d55c4a22dcbfc17d29
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22781
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot