nvmf: Add support for mDNS Pull Registration Requests (TP-8009, TP-8010a, and TP-8024)
To enable the SPDK NVMe-oF target discovery controller to act as a DDC, this patch added support for mDNS Pull Registration Requests. A DDC may respond to mDNS queries for the service name of "._nvme-disc._<protocol>.local." This patch covers the sending part of the mDNS message.

Change-Id: I3f759341c269a14d1a008e5577132880796d3d74
Signed-off-by: kyuho.son <kyuho.son@samsung.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23036
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>