test/mock: introduce DEFINE_WRAPPER_MOCK() macro
This macro is intended to be used to define the mock variables (e.g.
ut_foo_mocked and ut_foo) for functions wrapped with DECLARE_WRAPPER().
Currently, it's identical to DEFINE_RETURN_MOCK(), but we'll be changing
them in the next patch, so we'll need two separate macros.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I9d551f5c3f0659501e9937a25076818f4cbb62b1
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22810
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>