app/trace: fix for trace_print()
This patch fixes issue introduced in patch 22642.

If lcore is defined in trace_print() and initialized with
SPDK_TRACE_MAX_LCORE, the 'if' statement inside 'for'
loop is always true and trace size is printed for
every lcore. This is not intended use of this
function - if the user passes the lcore number
as a command line parameter, only this lcore shall
be printed. We need to propagate lcore number from
main(), where it is parsed and assigned, to the
trace_print().

Change-Id: I5c84da7da7df39c88ac52ba4ffaea8a518de77c6
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22829
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>