raid: write sb earlier when removing a base bdev
Update the superblock as soon as the raid bdev is quiesced, then
continue with removing the base bdev.

The condition in _raid_bdev_write_superblock() has to be changed because
base_info->desc is not NULL now, only remove_scheduled is set at this
stage of removing the base bdev.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Change-Id: I6a55d0ed5d0c1e3fc38825ac23d1556d42d2e801
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22842
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot