nvmf: make allow_any_host its own byte
This field can be modified from multiple threads, so it's protected by a
mutex.  However, it was defined as a single bit in a uin8t_t bitfield with
fields that could be accessed without taking that mutex, which could
lead to issues when multiple threads modified bits in that uint8_t.

Suggested-by: Jim Harris <jim.harris@samsung.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ib66e0b1f298455ff518c24895364269d3d0bb5f3
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22898
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>