lib/mlx5: Add native implementation of CQ and QP
Add functions to create, destory CQ and QP
Next patches add implementation of other API
used in IO path to replace mlx5_dv API. Under the
hood CQ and QP are created used verbs API and then
low level objects are extracted to be used directly.

mlx5_ifc.h is a binary interface to the HW according
to the PRM.
mlx5_ifc.h is a copy of rdma-core header version v51.0
with some modifications. Original copyright is kept.
This file is available under a choice of one of two
licenses: GPL v2 or BSD-2-Clause (the clause is not
specified explicitly but the text matches the 2-clause)
LICENSE file was updated with mentioning that mlx5_ifc.h
in SPDK tree is BSD-2-Clause

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: If5415729019c85eda8e6a1494164d3d0bab9ee94
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23102
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot