nvmf: disallow adding the same host to a subsystem many times
We used to allow adding the same host to the same subsystem many times.
However, this can cause issues, as we didn't check transport-specific
host options (e.g. TLS PSK), so adding the same host with different
transport options would succeed, which is obviously incorrect.

We could fix it by comparing host's options and ensuring they're the
same, but it's easier to simply disallow it altogether.  It's also
consistent with the behavior of other RPCs.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I1a60329100f5a3256c5ccc09db0586e7a023b6e5
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22896
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@intel.com>
Community-CI: Mellanox Build Bot