raid: keep raid bdev in CONFIGURING state when last base bdev is removed
When a raid bdev is created and there are none of the requested base
bdevs, it will be in CONFIGURING state with 0 discovered base bdevs,
waiting for the base bdevs to become available. So for consistency, when
a raid bdev is in CONFIGURING state with some discovered base bdevs,
don't delete the raid if those base bdevs are removed.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Change-Id: Ieb8e47b777b03a0a1194ed6e4b4713fa46e70940
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22932
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>