autotest: Run `sw_hotplug` test with ASan enabled
Orginal comment that prevent `sw_hotplug` execution with ASan clamis
that "... it catches SEGV earlier than our handler which breaks the
hotplug logic"

But when running ASan on all jobs
(https://review.spdk.io/gerrit/c/spdk/spdk/+/20722), the
above-mentioned problem did not appear. Even when ASan was enabled,
and the SPDK_RUN_ASAN flag was not set.

Therefore, it seems that ASan does not exclude the possibility of
running `sw_hotplug' anymore.

Change-Id: I3e371cf69d680fd9362c48d179f164f936102ea9
Signed-off-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23033
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>