test/scheduler: Meassure utime of $spdk_pid threads as a fallback
Normally these tests look at a business of given cpu as a whole but
this is a bit problematic when some external entity starts to steal
its time.  This causes discrepancies in tests where given cpu is
suddenly marked as busy where in fact SPDK was idle.

To mitigate this, after all per-cpu checks are done, and in case given
cpu is not idle where we expect it should be, check utime of the SPDK
thread pinned to that cpu.

Fixes issue #3362.

Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23074 (master)

(cherry picked from commit 7b52e4c17b1b1c2e5f4c92e8278e9736feb41dcc)
Change-Id: Ia105d1c45a9d0fa4c01a167a6c39790d68fb73ac
Signed-off-by: Marek Chomnicki <marek.chomnicki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23270
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>