scripts/pkgdep: Add support for shfmt-v3.8.0
This requires a little bit of a dance to keep using it under all
the distros we still support.

Introduce an _install_shfmt() wrapper which by default will try to
fetch release binary from github (default for all linux distro).
In second "build" mode, it will try to build shfmt under the go
version which our pkgdep supports - this is needed for FreeBSD since
upstream doesn't provide binaries for this distro anymore.

The end goal would be to squash everything into "build" mode but this
requires testing across all the Linux distros so it will be done in
later patches.

Change-Id: Ie2ff1a08e20b729814b5e8446e922425b5e5da7c
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23086
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot