scripts: Disable DSA and IAA default vfio-pci binding
The DSA and IAA device IDs have been added to the vfio denylist,
due to an erratum (see INTEL-SA-01084 security advisory).
This patch disables the default behavior of vfio-pci binding
for these devices.

To force the binding, it is possible to use the PCI_ALLOWED variable.

Change-Id: I25be317156f19f400f76961c5928d602043125dc
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23196
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot