test/nvme: use SPDK_CU_ASSERT_FATAL if nvme_pcie_poll_group_get_stats() fails
If it fails, tgroup_stat pointer does not get set, meaning the following
memcmp could cause a segfault.

This was found using LTO, even though the error message was somewhat confusing.
It reported there could be a source size of 0, not that one of the memcmp()
parameters could be NULL.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I6e72e0932c0cc2f701d90fdbfaac17adb8eb771c
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23307
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>