nvmf: fix deprecated "transport" check in decode_rpc_listen_address
The current decoder would always emit the deprecation notice, because even
if user specified 'trtype', it would get stored in ->transport and trigger
the warning.

While here, adjust some of the ERRLOGs to mention 'trtype' instead of
'transport' since that's the name we are using now.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I6b712e5050ff8d3603faace54e6e0f9ef7d88777
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23361
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>