event: deinit old scheduler before init'ing new one
I think the idea here was to try to init the new scheduler first,
and then if that fails we just keep the old scheduler instead.

But this breaks the common case where you want to switch between
the dynamic and gscheduler schedulers. For example, if you switch
from dynamic to gscheduler, gscheduler will init() and set the
governor, but then dynamic deinit() will clear the governor.
gscheduler() then segfaults because the governor is now NULL.

Fixes issue #3388.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I5ca031b7a0e8513c1e12a13e6989c57515f27358
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23382
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot