env: add spdk_env_core_get_smt_cpuset
This create a cpuset with core bits set for the specified core and
any of its SMT siblings.

If the core has siblings, then all of its siblings will be set, even
those that are not part of the application's core mask.

If the specified core is UINT32_MAX, then bits will be set for all
SMT siblings of all cores in the application's core mask.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I6e9677414fd3039089c4faec3464cf135a70a6ce
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23406
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>