nvmf: check oncs when executing IO commands
Transports are allowed to modify cdata including limiting advertised
support for optional IO commands.  So, commands that aren't marked as
supported in oncs will now be rejected with the INVALID_OPCODE status.

Additionally, this patch also sets the dsm and write_zeroes bits in
ctrlr->cdata.oncs.  That's now necessary, because previously the support
for these commands were only determined based on whether any of the
namespaces supported them.  Now, even when a namespace support a
command, a transport can still disallow it by clearing it in cdata.oncs.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I44f5c04bc01c37f18bd519dbaf2edfa0ab774eae
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23374
Reviewed-by: Jacek Kalwas <jacek.kalwas@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>